Rethink Yamux terminology and addition of muxers to the node
