--- core/libp2p/muxer/muxer_config.hpp
@@ -1,13 +1,0 @@
-/**
- * Copyright Soramitsu Co., Ltd. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#ifndef KAGOME_MUXER_CONFIG_HPP
-#define KAGOME_MUXER_CONFIG_HPP
-
-namespace libp2p::muxer {
-  struct MuxerConfig {};
-}  // namespace libp2p::muxer
-
-#endif  // KAGOME_MUXER_CONFIG_HPP
