+++ core/libp2p/muxer/yamux/CMakeLists.txt
@@ -6,6 +6,8 @@
 add_library(yamux
     yamux.cpp
     yamux.hpp
+    yamuxed_connection.cpp
+    yamuxed_connection.hpp
     yamux_config.hpp
     yamux_frame.cpp
     yamux_frame.hpp
