+++ core/libp2p/muxer/yamux/yamux_config.hpp
@@ -6,13 +6,11 @@
 #ifndef KAGOME_YAMUX_CONFIG_HPP
 #define KAGOME_YAMUX_CONFIG_HPP
 
-#include "libp2p/muxer/muxer_config.hpp"
-
 namespace libp2p::muxer {
   /**
    * Configuration used to create instance of Yamux multiplexer
    */
-  struct YamuxConfig : public MuxerConfig {
+  struct YamuxConfig {
     bool is_server;
   };
 }  // namespace libp2p::muxer
