+++ core/libp2p/muxer/yamux/yamux_frame.hpp
@@ -8,14 +8,14 @@
 
 #include <gsl/span>
 #include "common/buffer.hpp"
-#include "libp2p/muxer/yamux/yamux.hpp"
+#include "libp2p/muxer/yamux/yamuxed_connection.hpp"
 
 namespace libp2p::muxer {
   /**
    * Header with optional data, which is sent and accepted with Yamux protocol
    */
   struct YamuxFrame {
-    using StreamId = Yamux::StreamId;
+    using StreamId = YamuxedConnection::StreamId;
     static constexpr uint32_t kHeaderLength = 12;
 
     enum class FrameType : uint8_t {
