--- core/libp2p/security/connection_encryptor.hpp
@@ -1,13 +1,0 @@
-/**
- * Copyright Soramitsu Co., Ltd. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#ifndef KAGOME_ENCRYPTED_CONNECTION_HPP
-#define KAGOME_ENCRYPTED_CONNECTION_HPP
-
-namespace libp2p::security {
-  class ConnectionEncryptor {};
-}  // namespace libp2p::security
-
-#endif  // KAGOME_ENCRYPTED_CONNECTION_HPP
