+++ core/libp2p/security/connection_security.hpp
@@ -1,0 +1,16 @@
+/**
+ * Copyright Soramitsu Co., Ltd. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef KAGOME_CONNECTION_SECURITY_HPP
+#define KAGOME_CONNECTION_SECURITY_HPP
+
+namespace libp2p::security {
+  /**
+   * Strategy to upgrade connections to secured
+   */
+  class ConnectionSecurity {};
+}  // namespace libp2p::security
+
+#endif  // KAGOME_CONNECTION_SECURITY_HPP
