+++ core/libp2p/upgrader/connection_upgrader.hpp
@@ -8,8 +8,8 @@
 
 #include <memory>
 
-#include "libp2p/muxer/yamux/yamux.hpp"
-#include "libp2p/muxer/stream_muxer.hpp"
+#include "libp2p/muxer/yamux/yamuxed_connection.hpp"
+#include "libp2p/transport/muxed_connection.hpp"
 
 namespace libp2p::upgrader {
   /**
@@ -31,7 +31,7 @@ namespace libp2p::upgrader {
    public:
     virtual ~ConnectionUpgrader() = default;
 
-    using NewStreamHandler = muxer::Yamux::NewStreamHandler;
+    using NewStreamHandler = muxer::YamuxedConnection::NewStreamHandler;
 
     /**
      * @brief upgrades connection to muxed
@@ -40,10 +40,9 @@ namespace libp2p::upgrader {
      * @param handler generic protocol handler
      * @return muxed connection instance
      */
-    virtual std::unique_ptr<muxer::StreamMuxer>
-    upgradeToMuxed(std::shared_ptr<transport::Connection> connection,
-                   MuxerOptions muxer_options,
-                   NewStreamHandler handler) const = 0;
+    virtual std::unique_ptr<transport::MuxedConnection> upgradeToMuxed(
+        std::shared_ptr<transport::Connection> connection,
+        MuxerOptions muxer_options, NewStreamHandler handler) const = 0;
 
     // TODO(yuraz): PRE-149 implement upgradeToSecure()
   };
